Reworked Serial Connection
Slider now has its own class in which it passes a serial connection back
to the UI. You can make your slider vibrate, it can read its current
value and moving the slider to a specified position is being worked on.
The program will pull the ID, Port name, and baudrate from the setup.