4/28 Update
Setup screen has a refresh button now that will update the dropdown of
the comports allowing you to find the port without closing the software
if something happened where it didn't find it originally.

The sliderName label is changed, but not set.

Sync Volume should now change the slider's position to match the current
master volume. When setting this, it seems that you need to do a full
cycle up and down in order for it to sync properly with the volume.